Backend fixes

+++ backend/src/routes/userFriendsRoute.ts
@@ -1,49 +1,64 @@
-import { Router, Request, Response } from "express";
+import { Request, Response, Router } from "express";
 import { Session } from "neo4j-driver";
-import driver from "../driver/driver.js";
-import User from "../models/User.js";
-import {
-  OkErrorResponse,
-  FriendsPageErrorResponse,
-  FriendRequestsPageErrorResponse,
-  FriendSuggestionsPageErrorResponse,
-} from "../types/userResponse.js";
-import { deleteFriend } from "../userFriends.js";
-import { declineFriendRequest } from "../userFriends.js";
-import { acceptFriendRequest } from "../userFriends.js";
-import { sendFriendRequest } from "../userFriends.js";
-import { getFriendSuggestionsCount } from "../userFriends.js";
-import { getFriendSuggestions } from "../userFriends.js";
-import { getFriendRequestsCount } from "../userFriends.js";
-import { getFriendRequests } from "../userFriends.js";
-import { getFriendsCount } from "../userFriends.js";
-import { getFriends } from "../userFriends.js";
-import { userNotFoundRes } from "./usersRoute.js";
+
+import driver from "../driver.js";
+import { formatError } from "../misc/formatError.js";
+import { authenticateToken, JWTRequest } from "../misc/jwt.js";
 import { Errors } from "../models/Response.js";
 import Page, { pageSchema } from "../models/routes/Page.js";
-import { formatError } from "../misc/formatError.js";
+import User from "../models/User.js";
+import {
+  AuthOkErrorResponse,
+  FriendRequestsPageErrorResponse,
+  FriendsPageErrorResponse,
+  FriendSuggestionsPageErrorResponse,
+  OkErrorResponse,
+} from "../types/userResponse.js";
+import {
+  acceptFriendRequest,
+  declineFriendRequest,
+  deleteFriend,
+  getFriendRequests,
+  getFriendRequestsCount,
+  getFriends,
+  getFriendsCount,
+  getFriendSuggestions,
+  getFriendSuggestionsCount,
+  sendFriendRequest,
+} from "../userFriends.js";
+import { getDbUser, getTokenDbUser } from "../users.js";
+import { userNotFoundRes } from "./usersRoute.js";
 
 const friendsRouter = Router();
 
 friendsRouter.get(
   "/:userId/friends",
-  async (req: Request, res: FriendsPageErrorResponse) => {
-    const userId = req.params.userId;
-
-    const pageParse = pageSchema.safeParse(req.query);
-    if (!pageParse.success) {
-      const errors = formatError(pageParse.error);
-      return res.status(400).json({ status: "error", errors });
-    }
-
-    const { page, maxUsers }: Page = pageParse.data;
-    const maxUsersBig = BigInt(maxUsers);
-
+  authenticateToken,
+  async (req: JWTRequest, res: FriendsPageErrorResponse) => {
     const session = driver.session();
     try {
+      const userId = req.params.userId;
+      const user = await getTokenDbUser(session, req.token!);
+
+      if (!user) {
+        return userNotFoundRes(res);
+      }
+
+      if (user.id != userId) {
+        return res.status(403).json({ status: "forbidden" });
+      }
+
+      const pageParse = pageSchema.safeParse(req.query);
+      if (!pageParse.success) {
+        const errors = formatError(pageParse.error);
+        return res.status(400).json({ status: "error", errors });
+      }
+
+      const { page, maxUsers }: Page = pageParse.data;
+      const maxUsersBig = BigInt(maxUsers);
+
       const friends = await getFriends(session, userId, page - 1, maxUsers);
       if (friends === null) {
-        console.log(friends);
         return userNotFoundRes(res);
       }
 
@@ -55,7 +70,6 @@ friendsRouter.get(
       const pageCount = Number(
         (friendsCount.toBigInt() + maxUsersBig - 1n) / maxUsersBig,
       );
-      console.log(pageCount);
       return res.json({ status: "ok", pageCount, friends });
     } catch (err) {
       console.log("Error:", err);
@@ -112,20 +126,31 @@ friendsRouter.get(
 
 friendsRouter.get(
   "/:userId/friend-suggestions",
-  async (req: Request, res: FriendSuggestionsPageErrorResponse) => {
+  authenticateToken,
+  async (req: JWTRequest, res: FriendSuggestionsPageErrorResponse) => {
     const userId = req.params.userId;
-
-    const pageParse = pageSchema.safeParse(req.query);
-    if (!pageParse.success) {
-      const errors = formatError(pageParse.error);
-      return res.status(400).json({ status: "error", errors });
-    }
-
-    const { page, maxUsers }: Page = pageParse.data;
-    const maxUsersBig = BigInt(maxUsers);
 
     const session = driver.session();
     try {
+      const user = await getTokenDbUser(session, req.token!);
+
+      if (!user) {
+        return userNotFoundRes(res);
+      }
+
+      if (user.id != userId) {
+        return res.status(403).json({ status: "forbidden" });
+      }
+
+      const pageParse = pageSchema.safeParse(req.query);
+      if (!pageParse.success) {
+        const errors = formatError(pageParse.error);
+        return res.status(400).json({ status: "error", errors });
+      }
+
+      const { page, maxUsers }: Page = pageParse.data;
+      const maxUsersBig = BigInt(maxUsers);
+
       const friendSuggestions = await getFriendSuggestions(
         session,
         userId,
@@ -159,12 +184,23 @@ friendsRouter.get(
 
 friendsRouter.post(
   "/:userId1/send-friend-request/:userId2",
-  async (req: Request, res: OkErrorResponse) => {
+  authenticateToken,
+  async (req: JWTRequest, res: AuthOkErrorResponse) => {
     const session = driver.session();
     const userId1 = req.params.userId1;
     const userId2 = req.params.userId2;
 
     try {
+      const user = await getTokenDbUser(session, req.token!);
+
+      if (!user) {
+        return userNotFoundRes(res);
+      }
+
+      if (user.id != userId1) {
+        return res.status(403).json({ status: "forbidden" });
+      }
+
       const requestResult = await sendFriendRequest(session, userId1, userId2);
       if (!requestResult.success) {
         const { firstUserExists, secondUserExists } = requestResult;
@@ -193,12 +229,23 @@ friendsRouter.post(
 
 friendsRouter.post(
   "/:userId1/accept-friend-request/:userId2",
-  async (req: Request, res: OkErrorResponse) => {
+  authenticateToken,
+  async (req: JWTRequest, res: AuthOkErrorResponse) => {
     const session = driver.session();
     const userId1 = req.params.userId1;
     const userId2 = req.params.userId2;
 
     try {
+      const user = await getTokenDbUser(session, req.token!);
+
+      if (!user) {
+        return userNotFoundRes(res);
+      }
+
+      if (user.id != userId1) {
+        return res.status(403).json({ status: "forbidden" });
+      }
+
       const requestResult = await acceptFriendRequest(
         session,
         userId1,
@@ -244,12 +291,23 @@ friendsRouter.post(
 
 friendsRouter.post(
   "/:userId1/decline-friend-request/:userId2",
-  async (req: Request, res: OkErrorResponse) => {
+  authenticateToken,
+  async (req: JWTRequest, res: AuthOkErrorResponse) => {
     const session = driver.session();
     const userId1 = req.params.userId1;
     const userId2 = req.params.userId2;
 
     try {
+      const user = await getTokenDbUser(session, req.token!);
+
+      if (!user) {
+        return userNotFoundRes(res);
+      }
+
+      if (user.id != userId1) {
+        return res.status(403).json({ status: "forbidden" });
+      }
+
       const requestResult = await declineFriendRequest(
         session,
         userId1,
@@ -291,12 +349,23 @@ friendsRouter.post(
 
 friendsRouter.delete(
   "/:userId1/delete-friend/:userId2",
-  async (req: Request, res: OkErrorResponse) => {
+  authenticateToken,
+  async (req: JWTRequest, res: AuthOkErrorResponse) => {
     const session = driver.session();
     const userId1 = req.params.userId1;
     const userId2 = req.params.userId2;
 
     try {
+      const user = await getTokenDbUser(session, req.token!);
+
+      if (!user) {
+        return userNotFoundRes(res);
+      }
+
+      if (user.id != userId1) {
+        return res.status(403).json({ status: "forbidden" });
+      }
+
       const requestResult = await deleteFriend(session, userId1, userId2);
       if (!requestResult.success) {
         const { firstUserExists, secondUserExists, wasFriend } = requestResult;
